# openapi.yaml
openapi: 3.0.3
info:
  title: Order Service
  description: Demo microservice for managing customer orders.
  version: 1.0.0
servers:
  - url: http://localhost:4002
    description: Local dev server
tags:
  - name: orders
    description: Operations on orders
paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /orders:
    get:
      tags: [orders]
      summary: List all orders
      operationId: listOrders
      responses:
        '200':
          description: A list of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      tags: [orders]
      summary: Create a new order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInput'
      responses:
        '201':
          description: Created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /orders/{id}:
    get:
      tags: [orders]
      summary: Get an order by ID
      operationId: getOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The requested order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not found
    put:
      tags: [orders]
      summary: Update an order
      operationId: updateOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInput'
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not found
    delete:
      tags: [orders]
      summary: Delete an order
      operationId: deleteOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: string
          description: Unique order ID
        userId:
          type: string
          description: ID of the user who placed the order
        items:
          type: array
          description: List of item SKUs in the order
          items:
            type: string
        total:
          type: number
          description: Order total in USD
        status:
          type: string
          description: Order status (pending, shipped, delivered, cancelled)
      required:
        - id
        - userId
        - items
        - total
        - status
    OrderInput:
      type: object
      properties:
        id:
          type: string
          description: Unique order ID
        userId:
          type: string
          description: ID of the user who placed the order
        items:
          type: array
          description: List of item SKUs in the order
          items:
            type: string
        total:
          type: number
          description: Order total in USD
        status:
          type: string
          description: Order status (pending, shipped, delivered, cancelled)